% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nCells.R
\name{nCells}
\alias{nCells}
\title{Counting of cells that express a gene/feature}
\usage{
nCells(counts)
}
\arguments{
\item{counts}{A count matrix; more options TBA}
}
\description{
Takes a count matirx and returns feature metadata
}
