% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cell_types.R
\name{plot_cell_types}
\alias{plot_cell_types}
\title{Convenient cell composition plotting}
\usage{
plot_cell_types(object, idents = NULL, reduction = "umap")
}
\arguments{
\item{object}{A Seurat object; more input options TBA}

\item{reduction}{A metadata column name to group by}
}
\description{
Takes a Seurat object and an annotation column name
}
